+++ tests/testthat/helpers/testers.R
@@ -70,6 +70,8 @@ extfam_tester <- function(extfam,
   # TODO: Add some mathematical checks (i.e., check that the calculations for
   # the objects listed in `extfam_nms_add` are mathematically correct).
 
+  # Output ------------------------------------------------------------------
+
   return(invisible(TRUE))
 }
 
@@ -593,6 +595,15 @@ submodl_tester <- function(
           expect_length(submodl_totest[[!!j]]$w, nobsv)
         }
         expect_true(all(submodl_totest[[!!j]]$w > 0), info = info_str)
+        if (sub_fam == "gaussian") {
+          # Note: For non-Gaussian families, a comparison of
+          # `submodl_totest[[j]]$w` with `wobs_expected` doesn't make sense
+          # since `glm_ridge(<...>)$w` contains the weights of the
+          # pseudo-Gaussian observations as calculated in pseudo_data().
+          expect_equal(as.vector(submodl_totest[[!!j]]$w),
+                       wobs_expected %||% rep(1, nobsv),
+                       info = info_str)
+        }
 
         expect_s3_class(submodl_totest[[!!j]]$formula, "formula")
         if (!grepl(":", as.character(submodl_totest[[j]]$formula)[3])) {
@@ -860,13 +871,14 @@ projection_tester <- function(p,
   })
   sub_data_crr <- p$refmodel$fetch_data()
   if (p_type_expected) {
-    clust_ref <- .get_refdist(p$refmodel,
-                              nclusters = nprjdraws_expected,
-                              seed = seed_expected)
+    if (exists(".Random.seed", envir = .GlobalEnv)) {
+      rng_state_old <- get(".Random.seed", envir = .GlobalEnv)
+      on.exit(assign(".Random.seed", rng_state_old, envir = .GlobalEnv))
+    }
+    set.seed(seed_expected)
+    clust_ref <- .get_refdist(p$refmodel, nclusters = nprjdraws_expected)
   } else {
-    clust_ref <- .get_refdist(p$refmodel,
-                              ndraws = nprjdraws_expected,
-                              seed = seed_expected)
+    clust_ref <- .get_refdist(p$refmodel, ndraws = nprjdraws_expected)
   }
   for (i in seq_len(nprjdraws_expected)) {
     sub_data_crr[[y_nms[i]]] <- clust_ref$mu[, i]
@@ -1134,10 +1146,14 @@ vsel_tester <- function(
   expect_type(vs$search_path$submodls, "list")
   expect_length(vs$search_path$submodls, solterms_len_expected + 1)
   from_vsel_L1_search <- method_expected == "l1"
+  if (exists(".Random.seed", envir = .GlobalEnv)) {
+    rng_state_old <- get(".Random.seed", envir = .GlobalEnv)
+    on.exit(assign(".Random.seed", rng_state_old, envir = .GlobalEnv))
+  }
+  set.seed(seed_expected)
   clust_ref <- .get_refdist(vs$refmodel,
                             ndraws = ndraws_expected,
-                            nclusters = nclusters_expected,
-                            seed = seed_expected)
+                            nclusters = nclusters_expected)
   nprjdraws_expected <- ncol(clust_ref$mu)
   if (!from_vsel_L1_search) {
     y_nm <- as.character(vs$refmodel$formula)[2]