+++ tests/testthat/test_div_minimizer.R
@@ -1,6 +1,6 @@
 context("div_minimizer")
 
-test_that("`divmin` works", {
+test_that("divmin() works", {
   for (tstsetup in names(fits)) {
     args_fit_i <- args_fit[[tstsetup]]
     pkg_crr <- args_fit_i$pkg_nm
@@ -13,12 +13,12 @@ test_that("`divmin` works", {
       var_crr <- mean(as.matrix(fits[[tstsetup]])[, "sigma"]^2) +
         (colMeans(mu_crr^2) - colMeans(mu_crr)^2)
     } else {
-      var_crr <- rep(0, nobsv)
+      var_crr <- rep(NA, nobsv)
     }
     args_fit_i$projpred_var <- matrix(var_crr)
     args_fit_i$projpred_regul <- regul_default
 
-    if (args_fit_i$pkg_nm == "brms" && grepl("\\.with_wobs", tstsetup)) {
+    if (pkg_crr == "brms" && grepl("\\.with_wobs", tstsetup)) {
       args_fit_i$formula <- rm_addresp(args_fit_i$formula)
       args_fit_i$weights <- eval(args_fit_i$data)$wobs_col
     }