+++ tests/testthat/test_methods_vsel.R
@@ -119,7 +119,7 @@ test_that("`x` of class \"vselsummary\" (based on varsel()) works", {
         print(tstsetup)
         print(smmrys_vs[[tstsetup]], digits = 6)
       })
-      options(width = width_orig$width)
+      options(width_orig)
       if (testthat_ed_max2) local_edition(2)
     }
   }
@@ -141,7 +141,7 @@ test_that("`x` of class \"vselsummary\" (based on cv_varsel())  works", {
         print(tstsetup)
         print(smmrys_cvvs[[tstsetup]], digits = 6)
       })
-      options(width = width_orig$width)
+      options(width_orig)
       if (testthat_ed_max2) local_edition(2)
     }
   }
@@ -257,7 +257,7 @@ test_that("`nterms_max` is capped to the maximum model size", {
 context("suggest_size()")
 
 test_that("`stat` of invalid length fails", {
-  stopifnot(length(stats_common) > 0)
+  stopifnot(length(stats_common) > 1)
   skip_if_not(run_vs)
   for (tstsetup in head(names(vss), 1)) {
     expect_error(